Infinite Animated Parallax Background
Setup 3 background panels with 4 tiles sprites for animations. Set up with parallaxing effect which keeps it fairly self contained around the player, if the player exceeds 1300 from the leftmost panel it shifts to the front and should work indefinetly since new leftmost panel is like around 500 units from the player and it isn't visible at all so super juicy.