Cutscene & Tutorial Functional
Cutscene is setup to run kinda like pokemon cutscenes, typewriter text prints onto screen and you can click to fast forward the text, and click again to go to the next scene/frame. Fading is a bit buggy but idk we could use a dedicated "cover" transition instead of raw canvas groups.

Tutorial is functional, not the prettiest and i messed up the spacebar for input for desktop/pc players (thought it was left click) and it kind of is but idk. Tutorial seems a bit bland and un-informative, could use some work maybe. Buttons are also functional, not animating or the prettiest but they work and the dont show again method works nicely.

Dedicated cursor on Titlescreen (haven't tried honestly) and yeh, most stuff is disabled/enabled carefully on its own to work functionally rn.

CameraFollow bugs out if you disbale/enable it so we have a playerpref set to "pause" it i think, or an instance idk, one of the two