Butchery
So we gutted our MinimapManager and Planet/LevelManager to use pooling 😵 not all that performant increasing.

Also setup our planets themselves to be mesh quads instead of sprite renderers so they can be instanced, not really working either.

They do use a gpu instanced material tho which is better but still no visible changes.

Played with the spawn amount for planets/asteroids, does make a difference on the game feel and performance and the like, maybe over clutter so we dont travel as far as much even tho it might look like we have more objects existing idk.

Asteroids and currency gems need to be pooled/optimized since they drain a lot of physics cycles.