TitleScreenAudioPolish
Set on our 4 background songs and we will use a danger/warning "beep boop" in the background that will scale in speed based on distance from the blackhole to the player.

Got a MachRingExplosion.wav clip that we can scale the pitch based on the launch strength of the planet.

Button click is from ZeroOne <3 and we got a new blipHover.wav clip for Desktop users to have more "Ui interactivity" but idk about it 100% the button clicks are semi decent.

Toggles work entirely. Only setup for Titlescreen since we dont have much audio in our maingame and the prefabs are janked so no updating that. But yeh super pretty visuals and lerping / curves and allat. Simple AudioSource.mute can be used to mute audio and tracks continue in background.

TitleScreen needs Credits Panel and then its basically 100% done 😜