Finished Equipped Manager + Equip Text
Equip text changes from "equip" to "equipped" for the currently equipped ship type, works in preview menu aswell.

Equip Manager has entries for all 21 ships with animators, some animators are weird with their default states. Specter Tinyship13 was a bit weird and is pointing to move instead of idle, might have to fix the rest but for now its alright.