Shop Tabs + Trail Preview "Particles" + Cutscene Tweaks
So Shop tabs are functional, could use some more visual polish but pretty decent, slides both panels in sync to the visual slider between the buttons and changes the text color between them.

Preview UI thingy is a pool of just cube images to represent the trail, its not that great, functional, but not really what it looks like in game so idk, but a solid streak isnt all that visual either so idk.

Cutscene was played with, i changed the equippedmanager to disable the player not the cutscene manager since it was tweaking when rebuilding, it would just idk pulse weirdly with the timings but now it refreshes the first "idle" frame for the ship and immediately rebuilds the collider before disabling so its ready to go once the game starts. At least in theory thats how i hope it works. Had to gate it tho with our cutscene player pref key which kinda makes sense but yeh, tad touchy, seems good tho i ran through a few cutscenes and skip cutscene just re-runs and it holds up decently.

Bug***
Pausing while orbitting around a planet fixes our rotation and doesn't sync it with the orbit mechanics and whatnot, should be a simple like slap on the wrist type deal whenever we unpause the game. Pause Manager points to the orbit script or whatever