Improve Shop Selection Logic/Flow + Randomize Button Start
Improved clarity/flow between opening the shop and selected/equipped ship/trail info in the preview panel, it checks for any previously selected ships or trails and if it doesn't find any it uses the default/equipped keys.

Randomize button just calls a "select()" call for a random unlocked card and trail, works kinda okay but only equips the stuff in the designated tab not both unique/changed features.