V 1.0.1 - Any Size Screen Scaling
So our stupid defaults were set for a project that is meant to be dynamic and work in both portrait and landscape orientations. Expand or shink focuses better on one orientation so i switched everything to expand and it works beautifully.

Had to rearrange a few positions of the Credits Layout menu, but for the most part it should look good for any screen size.

Updated the Gem Collect() SFX to make it shorter, sounds a bit better