Tutorial Overhaul (Not Done)
Animated the planets to actually show you how to sling and release from the different input systems.

Fixed the bug where hitting asteroids on spawn or during tutorial would bork the rotation of the player, just made asteroids not spawn around the starting planet 👍

Fixed the tutorial so you can't use spacebar to escape the starting planet without closing the tutorial first

Fixed the cutscene so the player ship actually leaves the screen instead of getting stuck in the ceiling on different resolution sizes.